refactor(reconciler): extract org/service merge to a mixin — clears the §IX gate (P2 Slice 1)
merge_strategy.py was 879 lines (>600, unlisted §IX violation). Extract the
organization + service + field-provenance responsibility (merge_organization,
_merge_organization_data, merge_service, _merge_service_data, get_field_sources —
~442 lines) into app/reconciler/merge_org_service.py as OrgServiceMergeMixin, mixed
into MergeStrategy(OrgServiceMergeMixin, BaseReconciler). Methods are relocated
VERBATIM and still resolve self.db/self.logger (BaseReconciler) + self._row_to_dict
(MergeStrategy) on the composed instance — zero behaviour change, the public API
(MergeStrategy(db).merge_organization/.merge_service/.get_field_sources) is intact.

Result: merge_strategy.py 437 + merge_org_service.py 482 — both <600, §IX gate
cleared by extraction (no exception memo needed; location_creator.py is Slice 2).
283 reconciler characterization tests stay green. Also fixed a latent pre-existing
mypy [return-value] error (a bare `return` in merge_location's owner-protected
branch, confirmed on origin/main:368 — never caught because the reconciler is
excluded from the CI mypy gate, #559) by making it the consistent `return None`.
